requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
tinydb>=4.8.0
aiohttp>=3.8.0

